Refactor public tests to use describe.each
